Switched metadata file to csv format.
We were facing some troubles with updating .metadata.json so we switched the metadata file to CSV. Rest everything is the same.

There are 3 functions that you may find new:
1) metadata()
2) parentId()
3) countLines()

parentId() calculates the parent id for any row in .metadata.csv file.

countLines() calculates the number of rows in .metadata.csv file.

For some reason, the header line is not being written into the csv. Theres some logic error in the if statement which puts the header row if file is empty.
Create and check-in commands update it well. We still have to put the update in checkout.

We are pushing the most recent code on GitHub. We have modified only Main.java and CreateMain.java!!!